:wrench: [op,client] Prepare config files and logic.
Use 8401 as default RPC port.
Add OpHost and OpPort options to the client.
Add RPCHost and RPCPort options to the operator.
Expose GanacheCommand() to allow for BB testing.

+++ operator/operator.go
@@ -108,7 +108,7 @@ func Setup(cfg *Config) *Operator {
 }
 
 // Serve starts the operator's main routine.
-func (operator *Operator) Serve(port int) error {
+func (operator *Operator) Serve(port uint16) error {
 	// Handle errors, print them as they occurr
 	errg := perrors.NewGatherer()
 	errGo := func(name string, fn func() error) {
